consolidate production and dev navigation, adjust loader to wait on data in dev, limit how often we update json store to only scenarios that it needs to
